+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Update base text size in settings
+- Writer Persona ✍ : When writing to a file, specify 'pages' to a number in the tags and the llm will respond that many
+   times (capped at 10 for now)
 
 ### Changed